# openapi.yaml
openapi: 3.0.3
info:
  title: Cat API
  version: 1.0.0
  description: |
    Simple CRUD API for managing cats stored in MongoDB.

    Cats are identified by their MongoDB `_id` (a 24-character hex ObjectId string).
    When creating a cat, omit `_id` — MongoDB generates it automatically.

servers:
  - url: http://localhost:5000

paths:
  /cats:
    get:
      summary: List all cats
      operationId: getCats
      responses:
        "200":
          description: Array of all cats in the collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cat"
              example:
                - _id: "6627c4e2a1b2c3d4e5f60001"
                  name: Whiskers
                  age: 3
                  breed: Siamese
                  color: cream
                  indoor: true

    post:
      summary: Create a new cat
      operationId: createCat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatInput"
            example:
              name: Whiskers
              age: 3
              breed: Siamese
              color: cream
              indoor: true
      responses:
        "201":
          description: Cat created — response includes the MongoDB-generated _id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"

  /cats/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: MongoDB ObjectId (24-character hex string)
        schema:
          type: string
          pattern: "^[a-f0-9]{24}$"
          example: "6627c4e2a1b2c3d4e5f60001"

    get:
      summary: Get a cat by ID
      operationId: getCatById
      responses:
        "200":
          description: Cat found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
        "404":
          description: Cat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"

    put:
      summary: Update a cat (partial update supported)
      operationId: updateCat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatUpdate"
            example:
              age: 4
      responses:
        "200":
          description: Updated cat document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
        "404":
          description: Cat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"

    delete:
      summary: Delete a cat
      operationId: deleteCat
      responses:
        "204":
          description: Cat deleted — no response body
        "404":
          description: Cat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"

components:
  schemas:
    Cat:
      type: object
      required:
        - _id
        - name
        - age
        - breed
      properties:
        _id:
          type: string
          description: MongoDB ObjectId — 24-character hex string, generated by the database
          pattern: "^[a-f0-9]{24}$"
          example: "6627c4e2a1b2c3d4e5f60001"
        name:
          type: string
          description: The cat's name
          minLength: 1
          maxLength: 100
          example: Whiskers
        age:
          type: integer
          description: Age of the cat in years
          minimum: 0
          maximum: 30
          example: 3
        breed:
          type: string
          description: Breed of the cat
          minLength: 1
          maxLength: 100
          example: Siamese
        color:
          type: string
          description: Primary coat color
          example: cream
        indoor:
          type: boolean
          description: Whether the cat lives indoors
          example: true
        address:
          $ref: "#/components/schemas/Address"
      additionalProperties: false

    CatInput:
      type: object
      description: Fields to provide when creating or updating a cat. Do not include _id.
      required:
        - name
        - age
        - breed
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Whiskers
        age:
          type: integer
          minimum: 0
          maximum: 30
          example: 3
        breed:
          type: string
          minLength: 1
          maxLength: 100
          example: Siamese
        color:
          type: string
          example: cream
        indoor:
          type: boolean
          example: true
        address:
          $ref: "#/components/schemas/Address"
      additionalProperties: false

    CatUpdate:
      type: object
      description: Fields to provide when updating a cat (partial update). Do not include _id.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Whiskers
        age:
          type: integer
          minimum: 0
          maximum: 30
          example: 3
        breed:
          type: string
          minLength: 1
          maxLength: 100
          example: Siamese
        color:
          type: string
          example: cream
        indoor:
          type: boolean
          example: true
        address:
          $ref: "#/components/schemas/Address"
      additionalProperties: false
    Address:
      type: object
      required:
        - postalCode
        - country
      properties:
        postalCode:
          type: string
          example: "8011AB"
        country:
          type: string
          example: "NL"
      additionalProperties: false
    NotFound:
      type: object
      properties:
        message:
          type: string
          example: Cat not found